agents/pr-automation/github-linear-automation/requirements.txt
requests>=2.31.0
PyGithub>=2.1.1
GitPython>=3.1.40
python-dotenv>=1.0.0
pyyaml>=6.0.1
anthropic>=0.18.0
